requirements.txt
PyYAML==3.11
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.20
bottle==0.12.8
celery==3.1.18
distribute==0.6.24
eventlet==0.17.4
functools32==3.2.3-2
greenlet==0.4.7
jsonschema==2.5.1
kombu==3.0.26
pytz==2015.4
